# services/orchestrator-go/api/openapi.yaml
openapi: 3.0.3
info:
  title: MentatLab Orchestrator API
  description: |
    RESTful API for the MentatLab orchestrator service.

    ## Overview
    The orchestrator manages agent workflow execution as DAGs (Directed Acyclic Graphs).
    It provides endpoints for:
    - **Runs**: Create, start, cancel, and monitor workflow executions
    - **Agents**: Register and manage agent definitions
    - **Flows**: Store and retrieve reusable workflow definitions
    - **Artifacts**: Upload and download run artifacts
    - **Jobs**: Monitor Kubernetes job status (when running in K8s mode)

    ## Event Streaming
    Run events are streamed via Server-Sent Events (SSE) at `/api/v1/runs/{id}/events`.
    Events include logs, checkpoints, node status changes, progress updates, and control flow events.

    ## Authentication
    Currently no authentication is required. In production, requests should be
    proxied through the gateway which handles authentication.
  version: 1.0.0
  contact:
    name: MentatLab Team
  license:
    name: MIT
servers:
  - url: http://localhost:7070
    description: Local development
  - url: http://orchestrator:7070
    description: Kubernetes internal
...
paths:
...
  /api/v1/runs:
    post:
      tags: [runs]
      summary: Create a new run
      description: Creates a new run with an optional execution plan. Use `auto_start` to immediately begin execution.
      operationId: createRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRunRequest'
      responses:
        '201':
          description: Run created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRunResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags: [runs]
      summary: List runs
      description: Returns a paginated list of run IDs
      operationId: listRuns
      parameters:
        - name: limit
          in: query
          description: Maximum number of results (max 500)
          schema:
            type: integer
            default: 50
            maximum: 500
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of runs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListRunsResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/runs/{id}/cancel:
    post:
      tags: [runs]
      summary: Cancel a running execution
      operationId: cancelRun
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: Run cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: cancelled
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/jobs/{id}/status:
    get:
      tags: [jobs]
      summary: Get Kubernetes job status
      operationId: getJobStatus
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/jobs/{id}:
    delete:
      tags: [jobs]
      summary: Delete a Kubernetes job
      operationId: deleteJob
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '204':
          description: Job deleted
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

  # Diagnostics
...
  /api/v1/runstore/selfcheck:
    get:
      tags: [diagnostics]
      summary: RunStore self-check
      description: Creates a test run, appends an event, reads it back, and deletes it
      operationId: runStoreSelfCheck
      responses:
        '200':
          description: Self-check passed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  latency_ms:
                    type: integer
                  event_count:
                    type: integer
        '500':
          $ref: '#/components/responses/InternalError'
components:
  parameters:
    RunId:
      name: id
      in: path
      required: true
      description: Run ID
      schema:
        type: string
...
    JobId:
      name: id
      in: path
      required: true
      description: Kubernetes job ID
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: Service or dependency unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable message
        details:
          type: object
          additionalProperties: true

    # Run types
...
    Plan:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/NodeSpec'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/EdgeSpec'
    NodeSpec:
      type: object
      required: [id, type]
      properties:
        id:
          type: string
        type:
          type: string
          description: Node type (agent, conditional, for_each, subflow)
        agent_id:
          type: string
        command:
          type: array
          items:
            type: string
        image:
          type: string
        env:
          type: object
          additionalProperties:
            type: string
        inputs:
          type: array
          items:
            type: string
          description: IDs of upstream nodes
        timeout:
          type: integer
          description: Timeout in nanoseconds (Go duration)
        retries:
          type: integer
        conditional:
          $ref: '#/components/schemas/ConditionalConfig'
        for_each:
          $ref: '#/components/schemas/ForEachConfig'
        subflow:
          $ref: '#/components/schemas/SubflowConfig'
    EdgeSpec:
      type: object
      required: [from, to]
      properties:
        from:
          type: string
        to:
          type: string
    ConditionalConfig:
      type: object
      properties:
        expression:
          type: string
        then_node:
          type: string
        else_node:
          type: string
    ForEachConfig:
      type: object
      properties:
        collection:
          type: string
        item_var:
          type: string
        body_node:
          type: string
        max_parallel:
          type: integer
    SubflowConfig:
      type: object
      properties:
        flow_id:
          type: string
        inputs:
          type: object
          additionalProperties: true
    CreateRunRequest:
      type: object
      properties:
        name:
          type: string
        plan:
          $ref: '#/components/schemas/Plan'
        auto_start:
          type: boolean
          description: Start execution immediately after creation
    CreateRunResponse:
      type: object
      properties:
        runId:
          type: string
        status:
          type: string
        sse_url:
          type: string
          description: URL for SSE event stream (when auto_start is true)
...
    ListRunsResponse:
      type: object
      properties:
        runs:
          type: array
          items:
            type: string
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    # Event types
...
    JobStatus:
      type: object
      properties:
        job_id:
          type: string
        status:
          type: string
          enum: [pending, running, succeeded, failed, unknown]
        active:
          type: integer
        succeeded:
          type: integer
        failed:
          type: integer
        start_time:
          type: string
          format: date-time
        completion_time:
          type: string
          format: date-time